refactor(37): AABB into mesh header, meta kind uint64, builder crash-on-error
- Move AABB from metadata system into NtMeshAssetHeader (aabb_min/max
  float[3] fields, 24→48 bytes, version 1→2)
- Remove NtAabbData from pack format — AABB is inherent mesh data, not meta
- NtMetaEntryHeader.kind: uint32→uint64 (hash64, struct 16→20 bytes)
- Builder finish_pack: crash immediately on first import failure via
  NT_BUILD_ASSERT instead of accumulating fail_count
- Builder finish_pack: assert on NULL ctx and zero pending_count
- Remove tests that expected error returns from finish_pack (builder now
  asserts — bad input is not a recoverable error)
- Update all meta test helpers to use uint64_t kind and generic payloads

+++ engine/resource/nt_resource.c
@@ -890,7 +890,7 @@ const uint8_t *nt_resource_get_blob(nt_resource_t handle, uint32_t *out_size) {
 
 /* ---- Metadata query ---- */
 
-const void *nt_resource_get_meta(nt_resource_t handle, uint32_t kind, uint32_t *out_size) {
+const void *nt_resource_get_meta(nt_resource_t handle, uint64_t kind, uint32_t *out_size) {
     if (out_size) {
         *out_size = 0;
     }
